added new endpoints and fixed bug in getting all tables introspect endpoint
